tsd-backend-main/requirements.txt
# Core
Django==5.2.3
djangorestframework==3.16.0
mysqlclient==2.2.7
python-dotenv~=1.2.1

# API
drf-spectacular==0.28.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-cors-headers~=4.7.0
mozilla-django-oidc~=5.0.2
ldap3~=2.9.1

asgiref==3.8.1
attrs==25.3.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
PyJWT~=2.10.1
django-filter~=25.1
tabulate~=0.9.0
